Move NewInstance to deprecated and simplify CompilePreload
- Move NewInstance from src/Compiler/ to src-deprecated/Compile/
- Update namespace to BEAR\Package\Compile
- Add @deprecated annotation as it's no longer used with Ray.Compiler
- Simplify CompilePreload to call injector->getInstance() directly instead of using NewInstance

NewInstance was previously used for two purposes:
1. Track unbound dependencies (removed with CompileDependencies)
2. Load resource classes for preload generation (now done directly)

This change eliminates unnecessary abstraction and improves coverage.
